Move object tests to separate module
